+++ package/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.2.0 (2024-10-10)
+
+### Breaking Changes
+
+- Function rename:
+    - `get_dir_by_target` => `get_dir_by_target_reverse`
+    - `get_dir_by_targets` => `get_dir_by_targets_reverse`
+- Old function name, new functionality:
+    - `get_dir_by_target`
+    - `get_dir_by_targets`
+
+### What's New
+
+- Add `get_project_root_directory` to handle the error manually.
+
 ## 0.1.1 (2024-09-27)
 
 ### What's Changed